refatoração completa da aplicação p/ utilizar o framework flask
