+++ .gitignore
@@ -11,6 +11,7 @@ npm-debug.log*
 node_modules/
 jspm_packages/
 flask_session/
+older_version/
 
 # Compilação
 dist/
