+++ app/templates/index.html
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Login</title>
+  <link rel="stylesheet"
+    href="{{ url_for('static', filename='css/index.css') }}">
+</head>
+
+<body>
+  <div class="login-container">
+    <div class="logo">
+      <img src="{{ url_for('static', filename='assets/logo.svg') }}"
+        alt="BLOCKsky Logo" class="logo-img">
+      <h1 class="logo-text"><span class="block">BLOCK</span><i>sky</i>
+      </h1>
+    </div>
+    <form class="login-form" id="login-form">
+      <div class="form-group">
+        <label for="service">Service</label>
+        <input type="text" id="service" value="bsky.social" disabled
+          class="disabled-input">
+      </div>
+      <div class="form-group">
+        <label for="username">Username | Handle @</label>
+        <input type="text" id="username" name="username"
+          placeholder="Exemplo: xys.bsky.social" aria-label="Username" required>
+      </div>
+      <div class="form-group">
+        <div class="label-wrapper">
+          <label for="password">APP Password</label>
+          <div class="tooltip">
+            <span class="tooltip-icon">?</span>
+            <span class="tooltip-content">
+              Para obter o App Password, acesse Settings > Privacy and security
+              > App Passwords > Add App Password
+            </span>
+          </div>
+        </div>
+        <input type="password" id="password"
+          placeholder="Sua senha única no BSKY" aria-label="Password" required>
+      </div>
+      <button type="button" id="login-button"
+        class="login-button">Entrar</button>
+    </form>
+  </div>
+  <footer class="footer">
+    <div class="footer-content">
+      © 2024. Desenvolvido por <a href="https://github.com/c-camila"
+        target="_blank">c-camila</a> e <a href="https://github.com/fzsdev"
+        target="_blank">fzsdev</a>
+    </div>
+  </footer>
+  <script src="{{ url_for('static', filename='js/script.js') }}"></script>
+</body>
+
+</html>
