--- back/bin/atproto
@@ -1,8 +1,0 @@
-#!/home/felladev/Projetos/bluesky-blocker/bin/python3
-# -*- coding: utf-8 -*-
-import re
-import sys
-from atproto_cli import atproto_cli
-if __name__ == '__main__':
-    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
-    sys.exit(atproto_cli())
