--- back/bin/python3.10
@@ -1,1 +1,0 @@
-python3
