--- front/home.js
@@ -1,4 +1,0 @@
-document.addEventListener()
-print(
-  "não é documento vazio!!!"
-)
